RF_09.001.04 | remove the ManageJenkinsTest class and move the tests to the UserTest class
